requirements.txt
pdfplumber
pytesseract
Pillow
PyMuPDF
re
